+++ mykingapp/Today Page/TodaySuperTableViewController.swift
@@ -135,7 +135,7 @@ class TodaySuperTableViewController: UITableViewController {
     func dateFunc(){
         let today = Date()
         let format = DateFormatter()
-        format.dateFormat = "EEEE    MMMM d"
+        format.dateFormat = "EEEE \nMMMM d"
         DateLabel.text = format.string(from: today).uppercased()
     }
     
@@ -214,54 +214,4 @@ class TodaySuperTableViewController: UITableViewController {
         // Pass the selected object to the new view controller.
     }
     */
-
-    //ALERT CONTROLLER LOGIN POP UP
-    //maybe encorporate faceID/finger print reading functionality...
-    override func viewDidAppear(_ animated: Bool) {
-        createAlertLogin()
-    }
-    
-    func createAlertLogin(){
-        var usernameTextField: UITextField?
-        var passwordTextField: UITextField?
-        
-        let alertController = UIAlertController(
-            title: "Sign in",
-            message: "Please enter your credentials",
-            preferredStyle: .alert)
-
-        let loginAction = UIAlertAction(
-        title: "Log in", style: .default) {
-            (action) -> Void in
-            
-            if let username = usernameTextField?.text {
-                print(" Username = \(username)")
-            } else {
-                print("No Username entered")
-            }
-            
-            if let password = passwordTextField?.text {
-                print("Password = \(password)")
-            } else {
-                print("No password entered")
-            }
-        }
-        
-        alertController.addTextField {
-            (txtUsername) -> Void in
-            usernameTextField = txtUsername
-            usernameTextField!.placeholder = "Username"
-        }
-        
-        alertController.addTextField {
-            (txtPassword) -> Void in
-            passwordTextField = txtPassword
-            passwordTextField?.isSecureTextEntry = true
-            passwordTextField?.placeholder = "Password"
-        }
-        
-        alertController.addAction(loginAction)
-        present(alertController, animated: true, completion: nil)
-    }
-    
 }
